applyCoupon 분리
